feat: unified export of g2plot chart mapping
