+++ packages/chart-advisor/src/auto-plot.ts
@@ -1,26 +1,13 @@
 import { AdvisorOptions, Advice, dataToSpecs, specToLibConfig } from './advice-pipeline';
 import EventEmitter from '@antv/event-emitter';
 import * as G2Plot from '@antv/g2plot';
-import { uuid, translate, createLayer, DEFAULT_FEEDBACK } from './util';
+import { uuid, createLayer, DEFAULT_FEEDBACK } from './util';
 
 export interface Configs {
   title?: string;
   theme?: string;
   description?: string;
   data: any[];
-}
-
-/**
- *
- * @param advice - 图表类型以及通道配置
- * @param data - 数据
- * @param configs - 配置
- */
-function getConfig(advice: Advice, { title, theme, description, data }: Configs): any {
-  const configs: any = { ...specToLibConfig(advice, 'G2Plot').configs };
-  if (title) configs.title = { visible: true, text: title };
-  if (description) configs.description = { visible: true, text: description };
-  return { ...configs, theme, data };
 }
 
 export interface AutoPlotOptions extends AdvisorOptions {
@@ -127,22 +114,29 @@ export class AutoPlot extends EventEmitter {
     this.current = index;
     const { type } = advices[index];
     const currentType = advices[current].type;
-    const { title, description, theme } = options;
-    const configs = getConfig(advices[index], { title, description, theme, data });
-    this.currentConfigs = configs;
-    this.type = type;
-    if (plot && type === currentType) {
-      plot.updateConfig(configs);
-    } else {
-      if (plot) plot.destroy();
-      console.log(' 🐛🐛🐛 type');
-      console.log(type);
-      // @ts-ignore
-      this.plot = new G2Plot[translate(type)](container, configs);
+    const {
+      //title, description,
+      theme,
+    } = options;
+    // const configs = getConfig(advices[index], { title, description, theme, data });
+    const libConfig = specToLibConfig(advices[index], 'G2Plot');
+    if (libConfig) {
+      const configs: any = { ...libConfig?.configs, data, theme };
+      this.currentConfigs = configs;
+      this.type = type;
+      if (plot && type === currentType) {
+        plot.updateConfig(configs);
+      } else {
+        if (plot) plot.destroy();
+        // console.log(' 🐛🐛🐛 type');
+        // console.log(type);
+        // @ts-ignore
+        this.plot = new G2Plot[libConfig.type](container, configs);
+      }
+      this.plot!.render();
+      // 出发事件
+      this.emit('change', [index]);
     }
-    this.plot!.render();
-    // 出发事件
-    this.emit('change', [index]);
   }
 
   destroy() {
